Cleaned up dirbuster
