+++ main.py
@@ -1,6 +1,39 @@
 import sys
 import os
-from dirbuster import verify_domain, run
+import requests
+from dirbuster import run
+
+def verify_domain(domain):
+	"""
+	Verifies if a domain can be reached over HTTP or HTTPS and
+	returns the full URL or None if the website is offline.
+
+	Args:
+		domain (string): The domain to verify
+
+	Returns:
+		String|None: URL of online website
+	"""
+
+	url = None
+
+	try:
+		# See if we can reach domain over HTTP or HTTPS
+		request = requests.get('http://'+domain)
+
+		if request.status_code == 301:
+			request = requests.get('https://'+domain)
+			url = 'https://'+domain
+		else: 
+			url = 'http://'+domain
+
+		# Only succeed if the status code is 200
+		if request.status_code != 200:
+			url = None
+	except:
+		print("Domain does not exist")
+	
+	return url
 
 def main():
 	print("Jelle's awesome web scanner!")
@@ -17,7 +50,7 @@ def main():
 	if not wordlist_exists:
 		sys.exit("Oh no, no file!")
 
-	results = run(url)
+	results = run(url, wordlist)
 	print(results)
